Move LogicalRowExpressions to SPI
